scripts_python/requirements.txt
google-generativeai
pydantic
python-dotenv
requests
beautifulsoup4
pandas
string-similarity